`Development`: Add application table for applicants - client side (#173)
* feat: first draft application table for applicants

* feat: badge, adapted button row, criteria

* chore: client style

* fix: openapi

* fix: client style

* fix: client style

* chore: client-test

* chore: server test

* feat: client-style

* feat: reload after delete or withdrawing

* chore: client style

* tests: client tests

* chore: client-test

* chore: client-test

* fix: fixed build failure introduced through merge error

* fix: removed applicantId from signature

* chore: client-test

* chore: openapi

* chore: openapi

* chore: openapi

* chore: openapi

* fix: ckeaner constructor

* chore: openapi

* chore: openapi

* fix: add internal comments

* fix: create relative date in the server

* fix: nachtrag

* fix: removed debug comment

* fix: added comments, cleaned up code

* chore: openapi

* chore: clientstyle

* chore: clientstyle

* chore: architecture fix

---------

+++ src/main/java/de/tum/cit/aet/application/service/ApplicationService.java
@@ -3,12 +3,14 @@ package de.tum.cit.aet.application.service;
 import de.tum.cit.aet.application.constants.ApplicationState;
 import de.tum.cit.aet.application.domain.Application;
 import de.tum.cit.aet.application.domain.dto.ApplicationForApplicantDTO;
+import de.tum.cit.aet.application.domain.dto.ApplicationOverviewDTO;
 import de.tum.cit.aet.application.domain.dto.CreateApplicationDTO;
 import de.tum.cit.aet.application.domain.dto.UpdateApplicationDTO;
 import de.tum.cit.aet.application.repository.ApplicationRepository;
 import de.tum.cit.aet.core.constants.DocumentType;
 import de.tum.cit.aet.core.domain.Document;
 import de.tum.cit.aet.core.domain.DocumentDictionary;
+import de.tum.cit.aet.core.exception.EntityNotFoundException;
 import de.tum.cit.aet.core.exception.OperationNotAllowedException;
 import de.tum.cit.aet.core.service.DocumentDictionaryService;
 import de.tum.cit.aet.core.service.DocumentService;
@@ -203,7 +205,18 @@ public class ApplicationService {
      */
     @Transactional
     public void deleteApplication(UUID applicationId) {
+        if (!applicationRepository.existsById(applicationId)) {
+            throw new EntityNotFoundException("Application with ID " + applicationId + " not found");
+        }
         applicationRepository.deleteById(applicationId);
+    }
+
+    public List<ApplicationOverviewDTO> getAllApplications(UUID applicantId, int pageSize, int pageNumber) {
+        return applicationRepository.findApplicationsByApplicant(applicantId, pageNumber, pageSize);
+    }
+
+    public long getNumberOfTotalApplications(UUID applicantId) {
+        return this.applicationRepository.countByApplicant_UserId(applicantId);
     }
 
     /**
